DDP_improve
There is some problems when model.load_state_dict, because it saves model when train(DDP), but we only use single GPU to test. So I use technique to solve this problem. However, CUDA still out of memory. If want to solve this problem, maybe we also need use DDP to test.